+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 import AuthRouter from './modules/auth/routers/AuthRouter/AuthRouter';
-import ListRouter from './modules/auth/routers/ListRouter/ListRouter';
 import ProfileRouter from './modules/profile/routers/ProfileRouter/ProfileRouter.tsx';
 import CreateRouter from './modules/create/routers/CreateRouter/CreateRouter.tsx';
+import ListRouter from './modules/list/routers/ListRouter/ListRouter.tsx';
 
 const App: React.FC = () => {
   return (
