+++ src/modules/list/containers/QuizListContainer/QuizListContainer.css
@@ -1,3 +1,37 @@
-.quiz-list-container-content {
-    margin-left: 100px;
+.quiz-list-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 20px;
+  margin: 0 auto;
+  max-width: 500px;
+  width: 100%;
+  box-sizing: border-box;
+  background-color: #f8fafc;
+  border-radius: 15px;
+}
+
+.quiz-list-title {
+  font-size: 2.8rem;
+  font-weight: 1000;
+  color: #0056b3;
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+.quiz-list-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: var(--secondary-white);
+}
+
+.quiz-list {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  width: 100%;
 }
